requirements.txt
#  python==3.10

fastapi
fpdf2
google-generativeai
huggingface-hub
idna==3.4
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1719845458456/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1745673194/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1748333051527/work
keras==3.10.0
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lxml==5.4.0
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
networkx==3.3
numpy==1.26.4
pandas==2.2.3
pdfkit==1.0.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.0.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pydantic
PyMuPDF
pyparsing
python-docx
python-dotenv
python-multipart
PyYAML
regex
requests
spacy
streamlit
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow
tensorflow-io-gcs-filesystem==0.37.1
tf_keras==2.16.0
thinc==8.3.6
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
transformers
uvicorn